main 컴포넌트화
