+++ src/App.js
@@ -2,8 +2,8 @@ import styled from "styled-components";
 import Modal from "react-modal";
 import { getRandomNumber } from "./utils/numberUtils";
 import { getResult } from "./gamePolicy/getResult";
-import { useState, useEffect, useRef } from "react";
-import { Card } from "./Card";
+import { useState } from "react";
+import { Main } from "./Main";
 
 let randomNumber = getRandomNumber();
 let count = 1;
@@ -13,20 +13,6 @@ function App() {
   const [inputValues, setInputValues] = useState([]);
   const [cardListValues, setCardListValues] = useState([]);
   const [gamePlaying, setGamePlaying] = useState(false);
-
-  const bottomRef = useRef(null);
-
-  useEffect(() => {
-    bottomRef.current?.scrollIntoView({ behavior: "smooth" });
-  }, [cardListValues]);
-
-  function oninput(e) {
-    let number = Number(e.target.value);
-    if (e.target.value.length > 1) {
-      number = e.target.value[e.target.value.length - 1];
-    }
-    changeInputValue(number, inputFocus);
-  }
 
   function nextFocus() {
     if (inputFocus === 2) {
@@ -39,10 +25,6 @@ function App() {
     const clonedArray = [...inputValues];
     clonedArray[index] = number;
     setInputValues(clonedArray);
-  }
-
-  function changeInputfocus(e) {
-    setInputFocus(Number(e.target.id));
   }
 
   function confirmButtonClickHandler() {
@@ -60,6 +42,7 @@ function App() {
     randomNumber = getRandomNumber();
     count++;
   }
+
   return (
     <Wrapper>
       <ContentsWrapper>
@@ -78,29 +61,13 @@ function App() {
           <h1>제 {count}회 숫자 야구 게임</h1>
           <h1> {randomNumber}</h1>
         </Header>
-        <Main>
-          {/* 입력칸 3개 생성 */}
-          <InputWrapper>
-            {[0, 1, 2].map((id) => {
-              return (
-                <InputBox
-                  id={id}
-                  onFocus={changeInputfocus}
-                  type="number"
-                  onInput={oninput}
-                  value={inputValues[id]}
-                ></InputBox>
-              );
-            })}
-          </InputWrapper>
-          <CardList>
-            {/* 입력한 숫자에 대한 결과 카드 */}
-            {cardListValues.map((item) => {
-              return <Card item={item}></Card>;
-            })}
-            <div ref={bottomRef} />
-          </CardList>
-        </Main>
+        <Main
+          inputFocus={inputFocus}
+          changeInputValue={changeInputValue}
+          setInputFocus={setInputFocus}
+          inputValues={inputValues}
+          cardListValues={cardListValues}
+        />
         <Footer>
           {/* 숫자 버튼 */}
           <div>
@@ -192,12 +159,6 @@ const Header = styled.div`
   justify-content: center;
 `;
 
-const Main = styled.div`
-  width: 100%;
-  height: 400px;
-  //background-color: red;
-`;
-
 const Footer = styled.div`
   width: 100%;
   height: 160px;
@@ -225,49 +186,4 @@ const ConfirmButton = styled.button`
   //background-color: red;
   //color: #fff;
   padding: 10px;
-`;
-
-const InputWrapper = styled.div`
-  width: 100%;
-
-  display: flex;
-  justify-content: space-around;
-`;
-
-const InputBox = styled.input`
-  width: 50px;
-  height: 50px;
-  border: 1px solid black;
-  border-radius: 10px;
-  font-size: 50px;
-  text-align: center;
-  ::-webkit-outer-spin-button,
-  ::-webkit-inner-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-  }
-  :focus {
-    outline: none;
-  }
-`;
-
-const CardList = styled.div`
-  /* float: "left";
-  clear: "both"; */
-  display: grid;
-  width: 100%;
-  height: 300px;
-  //background-color: salmon;
-  grid-template-rows: 1fr 1fr;
-  grid-template-columns: 1fr 1fr 1fr;
-  overflow: overlay;
-  &::-webkit-scrollbar {
-    width: 5px;
-    border-radius: 3px;
-    background: rgba(255, 255, 255, 0.8);
-  }
-  &::-webkit-scrollbar-thumb {
-    background: rgba(0, 0, 0, 0.6);
-    border-radius: 3px;
-  }
 `;
